+++ src/com/sun/media/jsdt/http/SameVMSessionProxyThread.java
@@ -168,7 +168,7 @@ SameVMSessionProxyThread extends SameVMThread {
                 found = false;
                 synchronized (waitValueLock) {
                     if (waitValue != 0) {
-                        long s = (message.id        << 32) +
+                        long s = ((long) message.id << 32) +
                                  (message.sessionNo << 16) +
                                  (message.type      <<  8) + message.action;
