+++ src/com/sun/media/jsdt/http/SameVMThread.java
@@ -323,7 +323,7 @@ SameVMThread extends HttpThread {
                               " was: "        + waitValue);
                     }
 
-                    waitValue = (id   << 32) + (sessionNo << 16) +
+                    waitValue = ((long) id << 32) + (sessionNo << 16) +
                                 (type <<  8) + action;
 
                     if (SameVMThread_Debug) {
