+++ src/com/sun/media/jsdt/socket/SameVMThread.java
@@ -303,7 +303,7 @@ SameVMThread extends SocketThread {
                               " was: "        + waitValue);
                     }
 
-                    waitValue = (id   << 32) + (sessionNo << 16) +
+                    waitValue = ((long) id << 32) + (sessionNo << 16) +
                                 (type <<  8) + action;
 
                     if (SameVMThread_Debug) {
